WIP refactor default
