+++ lib/github-utils.js
@@ -1,0 +1,21 @@
+const git = require('simple-git')(process.cwd())
+
+const githubUrlRegex = /^git@github\.com:(.+)\.git$/
+
+export async function getRepo () {
+  return new Promise((resolve, reject) => {
+    git.getRemotes(true, (err, remotes) => {
+      if (err) return reject(err)
+
+      const origin = remotes.find(({ name }) => name === 'origin')
+      if (!origin) return reject('No origin specified on repo')
+
+      const decomposed = githubUrlRegex.exec(origin.refs.fetch)
+      resolve(decomposed[1])
+    })
+  })
+}
+
+export async function getVersion () {
+
+}
